solve ParseFieldData success test comparison issue
the problem was that the type in use is Dictionary<string, string[]> which uses a reference type, so a direct [Collection]Assert.AreEqual check will only compare references and return false.

i solved it by comparing the sizes of the expected and actual values, then iterating over the elements and comparing the keys with Assert.AreEqual and the values with CollectionAssert.AreEqual to force comparison of contents instead of references.